Revert "Revert "qssi : set vendor specific media profile""
This reverts commit d3099e0ae3ff60fc6c1f032ee438eb553dbc899a.

This change was originally added to solve some issues
related to camera profiles. But it created issue on
lahaina since a dependent change was missing. So the
original change was reverted and merged.

Now the dependent change got promoted to lahaina. So
making the change again to fix the original camera
profiles issue.

Change-Id: Ia77a17d0c0fd16acd1cd93a4ea0bfe98956f0cd7